requirements.txt
Django==5.1.4
djangorestframework==3.15.2
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
prometheus-client==0.21.1
python-json-logger==3.2.1
gunicorn==23.0.0
django-filter==24.3
